Sync MusyX source
Former-commit-id: 081a2fe39f957eabbe1ae649d9afc3aaf641e386

+++ src/musyx/runtime/CheapReverb/creverb.c
@@ -415,7 +415,7 @@ lbl_803B599C:
 #endif
 }
 #else
-static void HandleReverb(long* sptr, struct _SND_REVSTD_WORK* rv) {
+static void HandleReverb(s32* sptr, struct _SND_REVSTD_WORK* rv) {
   // TODO: Reimplement this in C
 }
 #endif
